New IR Instructions to support GC (#808)
* - Add another optimization variation + fixed typo

* - More Optimizations

* - More Optimizations

* - Progress

* - New IR instructions to separate object references (part of the GC plan)

+++ Source/Mosa.Utility.SourceCodeGenerator/TransformExpressions/Tokenizer.cs
@@ -192,7 +192,7 @@ namespace Mosa.Utility.SourceCodeGenerator.TransformExpressions
 			}
 
 			// integer
-			ulong l = ulong.Parse(part);
+			ulong l = (ulong)long.Parse(part);
 
 			return new Token(TokenType.IntegerConstant, index, part, l);
 		}
